Add files via upload
In this sales data have highlighted the importance of Advance filter in excel by using Sort & Filter to find sales greater than for 2 Sales person , Sub-Total to calculate the total commission earned by the sales associate. On commission worksheet created a Conditional Formatting method to apply an icon set to the values in the Rating column.